Create DataJobs_SalaryAnalysis_by_Sex
The project set out to perform an extensive analysis focusing on engineers, specifically various categories of data engineers like data analysts and data scientists. Key factors such as age, gender, country of residence, and professional experience were considered in this analysis.

To ensure the reliability of the findings, the project began with meticulous data cleaning procedures. This included tasks such as averaging relevant columns, splitting data where necessary, and eliminating redundant columns that didn't contribute to the analysis.

Following the data cleaning phase, the analysis proceeded to derive significant conclusions. Notably, a prominent aspect of the findings centered around the average salaries of male and female engineers across different regions worldwide. This insight into salary disparities based on gender and geographic location provides valuable information for understanding the dynamics within the engineering field.

For a more detailed exploration of the results and to ensure accurate interpretation, a link to access the comprehensive findings is provided. It's essential for users to have the Power BI app installed on their laptops to view the results effectively. This ensures that the insights gleaned from the analysis can be properly examined and utilized for informed decision-making.